Restructure resolution test packages
